Expose Alamofire.Response instead of NSHTTPURLResponse
There are valid cases where the callers might need access to the response data and just having the NSHTTPURLResponse isn't enough. Dominic described a valid argument for this here:

https://github.com/RobotsAndPencils/Astro/pull/15#discussion_r60831808

I haven't added any automatic way to parse the JSON into models in the error case but I think that could be done through an extension.