Extend Comparable with clamp and clamped methods
The existing free clamp function was useful, but didn't read as nicely as instance methods on Comparable do.

This change adds mutating and non-mutating instance methods. Instead of taking min and max values they now take ClosedRanges which encapsulate lower and upper bounds.

This would normally be a breaking change because of the removal of the free clamp function, but it turns out that it was never made public... So this ends up being purely additive.