Add Sequence extension with overloads taking KeyPaths
Sequence methods that take a mapping function or predicate are often used with a closure that is equivalent to a KeyPath.

The Swift forum contains a few discussions that are relevant:

- https://forums.swift.org/t/pitch-keypath-based-map-flatmap-filter/6266
- https://forums.swift.org/t/treating-a-type-as-a-function/14627/20

But, I want this syntax now :) That being said, we can try to implement this for ourselves with some foresight in order to keep things source compatible with what might be added to the standard library in the future.

This introduces overloads for the appropriate methods that take KeyPaths. A function called `get` is added that converts KeyPaths to their equivalent functions in order to reduce the boilerplate of inline closures. The spelling of `get` was borrowed from [swift-overture](https://github.com/pointfreeco/swift-overture/blob/e3e63876688ca8b6f327e5af7018f5e93e742367/Sources/Overture/KeyPath.swift#L9).